// frontend/src/app/modules/apps/customer/models/customer.model.ts
export class CustomerModel {
  firstName: string;
  middleName: string;
  lastName: string;
  street1: string;
  street2: string;
  zipCode: string;
  state: string;
  city: string;
  telephone: string;
  cellPhone: string;
  fax: string;
  email: string;
  privateNotes: string;
  status: boolean;
  creditLimit: number;
  isCreditHold: boolean;
  availableCredit: number;
  paymentTerms: number;
  mcNumber: number;
  mcNumberType: string;
  usdotNumber: number;
  weightUnit: string;
  distanceUnit: string;
  temperatureUnit: string;
}